requirements.txt
langchain
openai
cohere
faiss-cpu
python-dotenv
